+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/result/BaseGeocodingCompatSearchSuggestion.kt
@@ -29,7 +29,4 @@ data class BaseGeocodingCompatSearchSuggestion(
     @IgnoredOnParcel
     override val type: BaseSearchSuggestionType =
         BaseSearchSuggestionType.SearchResultSuggestion(listOf(searchResultType))
-
-    override val isBatchResolveSupported: Boolean
-        get() = true
 }