+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/tests_support/BlockingSearchSelectionCallback.kt
@@ -1,7 +1,6 @@
 package com.mapbox.search.tests_support
 
 import com.mapbox.search.ResponseInfo
-import com.mapbox.search.SearchMultipleSelectionCallback
 import com.mapbox.search.SearchSelectionCallback
 import com.mapbox.search.common.tests.BaseBlockingCallback
 import com.mapbox.search.result.SearchResult
@@ -9,8 +8,7 @@ import com.mapbox.search.result.SearchSuggestion
 
 internal class BlockingSearchSelectionCallback :
     BaseBlockingCallback<BlockingSearchSelectionCallback.SearchEngineResult>(),
-    SearchSelectionCallback,
-    SearchMultipleSelectionCallback {
+    SearchSelectionCallback {
 
     override fun onSuggestions(suggestions: List<SearchSuggestion>, responseInfo: ResponseInfo) {
         publishResult(SearchEngineResult.Suggestions(suggestions, responseInfo))
@@ -30,14 +28,6 @@ internal class BlockingSearchSelectionCallback :
         responseInfo: ResponseInfo
     ) {
         publishResult(SearchEngineResult.Results(results, responseInfo))
-    }
-
-    override fun onResult(
-        suggestions: List<SearchSuggestion>,
-        results: List<SearchResult>,
-        responseInfo: ResponseInfo
-    ) {
-        publishResult(SearchEngineResult.BatchResult(results, responseInfo))
     }
 
     sealed class SearchEngineResult {
@@ -64,7 +54,6 @@ internal class BlockingSearchSelectionCallback :
         data class Suggestions(val suggestions: List<SearchSuggestion>, val responseInfo: ResponseInfo) : SearchEngineResult()
         data class Result(val result: SearchResult, val responseInfo: ResponseInfo) : SearchEngineResult()
         data class Results(val results: List<SearchResult>, val responseInfo: ResponseInfo) : SearchEngineResult()
-        data class BatchResult(val results: List<SearchResult>, val responseInfo: ResponseInfo) : SearchEngineResult()
         data class Error(val e: Exception) : SearchEngineResult()
     }
 }